moved registration.js to pages
